opengl and serialization test updates (#40)
* unify header include style

* opengl related acf fixes and …

* support chnsCompute + isInit: see https://github.com/elucideye/acf/issues/37
* support rgb-to-gray w/ correct channel count==1 : see https://github.com/elucideye/acf/issues/36
* `*.mat` serialization testing
  + download Inria/Caltech *.mat toolbox pedestrian models from github storage using cmake/hunter_private_data
  + add simple serialization tests in test-acf GTest code (more rigorous testing needed)
  + remove unused code blocks and some debug code in test-acf.cpp
* fix a boundary case in sse gradHist() code
* remove unused conditional cv::cvtColor() code blocks in rgbConvert.cpp
* add virtual destructor to ObjectDetector.h base class
* support acf shrink (via constructor input) in the GPUACF code
* update gradhist glsl shader to highp precision
* unify header include style
* remove various comment blocks + format fixing
* add acf special member isLuv to indicate special case input LUV processing (such as in gpgpu pipeline)
* provide naive GLDetector : public acf::Detector naive detector to exercise shader code in acf-detect
  + use aglet lib for portable context
  + compute acf pyramids from input cv::Mat (in practice this would be separated for efficiency)

* clang-format

* remove logging

* remove output directory command line argument

use working directory

* test data download clarification

* test data updates

* tell user to update hunter if hunter_private_data is missing
* now… assume test data is available in gauze_add_test
* update drishti-upload submodule for newer hunter version so hunter_private_data is defined in ci builds

* debug ci failure to read test image on some platforms

* remove ci debugging

* use PNG format test image

Some of the platforms in drishti-upload are built without JPEG in OpenCV, so we use PNG format for now.  The OpenCV CI configs can probably be updated and unified.

+++ src/lib/acf/gpu/gradhist.cpp
@@ -8,7 +8,7 @@
 
 */
 
-#include "acf/gpu/gradhist.h"
+#include <acf/gpu/gradhist.h>
 
 BEGIN_OGLES_GPGPU
 
@@ -26,7 +26,7 @@ BEGIN_OGLES_GPGPU
 const char * GradHistProc::fshaderGradHistSrcN = OG_TO_STR
 (
 #if defined(OGLES_GPGPU_OPENGLES)
- precision mediump float;
+ precision highp float;
 #endif
  varying vec2 vTexCoord;
  uniform sampler2D uInputTex;
